made removeHandler public and moved SelectOptions to the correct folder and refactored its import statements
