+++ src/SquareEdit/Grouping.js
@@ -3,7 +3,7 @@
 import * as Contents from './Contents.js';
 import * as Warnings from '../Warnings.js';
 import * as Notification from '../utils/Notification.js';
-import { unsetVirgaAction, unsetInclinatumAction } from '../utils/SelectOptions.js';
+import { unsetVirgaAction, unsetInclinatumAction } from './SelectOptions.js';
 const $ = require('jquery');
 
 /**
